** Fixes bug #331099 based on patches by Eric Anholt         <anholt@FreeBSD.org> and Joe Marcus Clarke <marcus@FreeBSD.org>
        * camel-object.c: (cobject_state_read): Cast the input for
        g_try_malloc to gulong.

+++ camel/camel-object.c
@@ -457,7 +457,7 @@ cobject_state_read(CamelObject *obj, FILE *fp)
 		}
 		
 		/* we batch up the properties and set them in one go */
-		if (!(argv = g_try_malloc (sizeof (*argv) + (count - CAMEL_ARGV_MAX) * sizeof (argv->argv[0]))))
+		if (!(argv = g_try_malloc ((gulong)(sizeof (*argv) + (count - CAMEL_ARGV_MAX) * sizeof (argv->argv[0])))))
 			return -1;
 		
 		argv->argc = 0;
